+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Stac 1.0.0-rc.4 compatible release
 
 - Raster extension
 - Version extension
+- Virtual Assets extension
 - Processing extension completed
 
 ## Fixed